Moved tiles to new folder, created tile test adn empty tile strategy.
