--- boardgames/src/main/java/no/ntnu/idatx2003/exam2025/boardgames/model/TileStrategy.java
@@ -1,5 +1,0 @@
-package no.ntnu.idatx2003.exam2025.boardgames.model;
-
-public interface TileStrategy {
-  void applyEffect();
-}
