+++ boardgames/src/main/java/no/ntnu/idatx2003/exam2025/boardgames/model/tile/EmptyTileStrategy.java
@@ -1,0 +1,4 @@
+package no.ntnu.idatx2003.exam2025.boardgames.model.tile;
+
+public class EmptyTileStrategy {
+}
