+++ boardgames/src/main/java/no/ntnu/idatx2003/exam2025/boardgames/model/tile/TileStrategy.java
@@ -1,0 +1,5 @@
+package no.ntnu.idatx2003.exam2025.boardgames.model.tile;
+
+public interface TileStrategy {
+  void applyEffect();
+}
