+++ boardgames/src/test/java/no/ntnu/idatx2003/exam2025/boardgames/model/TileTest.java
@@ -1,0 +1,29 @@
+package no.ntnu.idatx2003.exam2025.boardgames.model;
+
+import no.ntnu.idatx2003.exam2025.boardgames.model.tile.Tile;
+import org.junit.jupiter.api.Test;
+
+class TileTest {
+
+  Tile tile = new Tile(5);
+
+  @Test
+  void getId() {
+  }
+
+  @Test
+  void setId() {
+  }
+
+  @Test
+  void getTileStrategy() {
+  }
+
+  @Test
+  void setTileStrategy() {
+  }
+
+  @Test
+  void applyTileEffect() {
+  }
+}
